Add golint and separate out loadbalancer interface
