+++ Makefile
@@ -1,10 +1,11 @@
 pkgs = $(shell go list ./... | grep -v /vendor/)
+files = $(shell find . -path ./vendor -prune -o -name '*.go' -print)
 
-all : format vet test build
+all : format vet lint test build
 
 format :
 	@echo "== formatting"
-	goimports -w $(shell find . -path ./vendor -prune -o -name '*.go' -print)
+	goimports -w $(files)
 
 test :
 	@echo "== running tests"
@@ -18,4 +19,10 @@ vet :
 	@echo "== vetting code"
 	go vet $(pkgs)
 
-.PHONY: all format test build vet
+lint :
+	@echo "== linting code"
+	@for pkg in $(pkgs); do \
+		golint -set_exit_status $$pkg; \
+	done;
+
+.PHONY: all format test build vet lint
